fix: handle device going in an out of range
services can vanish between connections if the device is
on the edge of bluetooth range. In this case we need
to avoid using the cache. This isnt a problem in the
native bleak implementation but its not merged yet.

This workaround is wrapped with a broad exception
catch so it assumes services cannot be cached if
anything fails in case bleak changes something
under the hood since its making some assumptions
about the implemenation that will not be safe in
future versions